Adds Favourite Contact Groups. Allows adding of contacts to Favourites.
Check if the group exists, creates if it doesn't, publishes the group for VM/Context to utilise.

Does not currently check if Contact is already in Group however doesn't seem to cause an issue. Currently no way outside
of the contacts app to remove a favourite. Also no indicator that a contact is already a favourite when they are
displayed in the list.